more plotting

+++ src/demo.m
@@ -1,4 +1,4 @@
-function demo(train, varargin)
+function [results, num_windows] = demo(train, varargin)
 %DEMO Demo implementation
 %
 %   Syntax: demo(train, ...)
@@ -18,6 +18,7 @@ function demo(train, varargin)
     params = get_default_configuration;
     params.log_file = 'demo.log';
     params.log_level = 'debug';
+    params.no_create = false;
     keywords = parse_keywords(varargin, fieldnames(params));
     params = merge_structs(params, keywords);
     params.default_bounding_box = params.get_bounding_box(params);
@@ -25,7 +26,7 @@ function demo(train, varargin)
     log_file(params.log_file);
     set_log_level(params.log_level);
     % print params
-    debg('Params:\n%s', struct2str(params));
+    info('Params:\n%s', struct2str(params));
 
     %params = profile_start(params);
 
@@ -45,7 +46,7 @@ function demo(train, varargin)
         succ('DONE');
     else
         cluster_model = generateCluster(params);
-        searchInteractive(params, cluster_model);
+        [results, num_windows] = searchInteractive(params, cluster_model);
     end
 end
 
@@ -112,7 +113,7 @@ function svm_models = getSVM(params, cluster_model)
     svm_models = get_svms(params, query_codebooks, neg_codebooks);
 end
 
-function results = searchInteractive(params, cluster_model)
+function [results, num_windows] = searchInteractive(params, cluster_model)
 %SEARCHINTERACTIVE Do a complete interactive database search
 %   Asks the user for a image in the dataset image path.
 %   Allows to mark a ROI to search for
@@ -130,7 +131,7 @@ function results = searchInteractive(params, cluster_model)
     start_time = tic;
 
     if usejava('desktop')
-        neg_codebooks_job = parfeval(@get_neg_codebooks, 1, params);
+%        neg_codebooks_job = parfeval(@get_neg_codebooks, 2, params, cluster_model);
         f = figure('Position', [100, 100, 1024+40, 768 + 100], 'Resize', 'off');
         axis image;
         ax = axes('Units', 'pixels', 'Position', [20, 80, 1024, 768]);
@@ -181,10 +182,11 @@ function results = searchInteractive(params, cluster_model)
 
     % run directly if in cli mode
     if ~usejava('desktop')
-        results = process([], []);
+        [results, num_windows] = process([], []);
     else
         uiwait;
-        results = btn.UserData;
+        results = btn.UserData{1};
+        num_windows = btn.UserData{2};
     end
 
     function setStatus(txt)
@@ -229,7 +231,7 @@ function results = searchInteractive(params, cluster_model)
         shg;
     end
 
-    function results = process(source, cbdata)
+    function [results, num_windows] = process(source, cbdata)
         profile_log(params);
         if usejava('desktop')
             selected_img = img.UserData;
@@ -247,11 +249,14 @@ function results = searchInteractive(params, cluster_model)
         [~, curid, ~] = fileparts(selected_img.filename);
 
         target_dir = get_target_dir(params, curid);
-        if exist(target_dir, 'dir')
-            rmdir(target_dir, 's');
+
+        if ~params.no_create
+            if exist(target_dir, 'dir')
+                rmdir(target_dir, 's');
+            end
+            debg('++ Using target dir %s', target_dir);
+            mkdir(target_dir);
         end
-        debg('++ Using target dir %s', target_dir);
-        mkdir(target_dir);
 
         if usejava('desktop')
             pos = round(getPosition(h));
@@ -260,8 +265,10 @@ function results = searchInteractive(params, cluster_model)
         roi_size = pos([3 4]);
         pos([3 4]) = pos([3 4]) + pos([1 2]) - 1;
 
-        croppedI = selected_img.I(pos(2):pos(4), pos(1):pos(3), :);
-        imwrite(croppedI, [target_dir filesep 'query.jpg']);
+        if ~params.no_create
+            croppedI = selected_img.I(pos(2):pos(4), pos(1):pos(3), :);
+            imwrite(croppedI, [target_dir filesep 'query.jpg']);
+        end
 
         if usejava('desktop')
             %database_job = parfeval(@load_database, 1, params, cluster_model, roi_size);
@@ -270,7 +277,7 @@ function results = searchInteractive(params, cluster_model)
         if exist('neg_codebooks_job', 'var')
             [~, neg_codebooks] = fetchNext(neg_codebooks_job);
         else
-            neg_codebooks = get_neg_codebooks(params);
+            neg_codebooks = get_neg_codebooks(params, cluster_model);
         end
 
         setStatus('Get features from selected part...');
@@ -337,23 +344,26 @@ function results = searchInteractive(params, cluster_model)
                 database = load_database(old_params, cluster_model, roi_size);
             end
             setStatus('Searching in database...');
-            results = searchDatabase(old_params, database, svm_models, fit_params, pos);
+            [results, num_windows] = searchDatabase(old_params, database, svm_models, fit_params, pos);
         else
             database = load_window_database(old_params, cluster_model, roi_size);
             setStatus('Searching in database...');
             results = searchWindowDatabase(old_params, database, svm_models, fit_params, pos);
+            num_windows = [];
         end
 
         % remove unnecessary fields
         cleanparams = clean_struct(params, {'cache', 'profile'});
 
         elapsed_time = toc(start_time);
-        save_ex([target_dir filesep 'results.mat'], 'results', 'cleanparams', 'elapsed_time');
+        if ~params.no_create
+            save_ex([target_dir filesep 'results.mat'], 'results', 'cleanparams', 'elapsed_time', 'num_windows');
+        end
         setStatus('DONE!');
         profile_stop(params);
 
         if usejava('desktop')
-            btn.UserData = results;
+            btn.UserData = {results, num_windows};
             uiresume;
         end
     end
@@ -370,7 +380,7 @@ function results = searchInteractive(params, cluster_model)
     end
 end
 
-function results = searchDatabase(params, database, svm_models, fit_params, pos)
+function [results, num_windows] = searchDatabase(params, database, svm_models, fit_params, pos)
 %SEARCHDATABASE Search the image database with the given SVM models
 %   Stores results in the results/queries subdirectories
 %
@@ -386,35 +396,8 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
 %   Output:
 %       results - Cell array of results. Fields: curid, img, patch, score
     profile_log(params);
-    sizes = {database.I_size};
-    sizes = cell2mat(vertcat(sizes(:)));
-    scale_factors = {database.scale_factor};
-    scale_factors = cell2mat(vertcat(scale_factors(:)));
-    max_w = max(sizes(:, 3) ./ scale_factors);
-    max_h = max(sizes(:, 4) ./ scale_factors);
 
-    roi_w = pos(3) - pos(1) + 1;
-    roi_h = pos(4) - pos(2) + 1;
-    windows = calc_windows(params, max_w, max_h, roi_w  * 0.75, roi_h * 0.75);
-    [ bboxes, codebooks, images ] = calc_codebooks(params, database, windows, params.parts, svm_models);
-    % save space
-    % database = rmfield(database, 'I');
-    % if isfield(database, 'tree')
-    %     database = rmfield(database, 'tree');
-    % end
-
-    % expand bounding boxes by 1/2 of patch average
-    if params.expand_bboxes
-        patch_avg = ceil((max(vertcat(database.max_size)) + min(vertcat(database.min_size))) / 2 / 2);
-        patch_avg = repmat(patch_avg, [size(bboxes, 1) 2]);
-        patch_avg = patch_avg .* repmat([-1 -1 1 1], [size(bboxes, 1) 1]);
-        bboxes = bboxes + patch_avg;
-        bboxes(:, [1 3]) = max(1, min(bboxes(:, [1 3]), max_w));
-        bboxes(:, [2 4]) = max(1, min(bboxes(:, [2 4]), max_h));
-    end
-
-    % required for libsvm
-    codebooks = double(codebooks);
+    [bboxes, codebooks, images, windows, num_windows] = extract_codebooks(params, svm_models, database, pos);
     results = cell([1 length(svm_models)]);
     for mi=1:length(svm_models)
         model = svm_models(mi);
@@ -442,7 +425,7 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
         mbbs = round(mbbs(idx, :));
         result = struct;
         if ~isempty(scores)
-            result = alloc_struct_array(length(scores), 'curid', 'query_curid', 'img', 'patch', 'score', 'bbox', 'filename');
+            result = alloc_struct_array(length(scores), 'curid', 'query_curid', 'img', 'patch', 'score', 'bbox', 'filename', 'num_windows');
             result(length(scores)).query_curid = model.curid;
         end
 
@@ -462,22 +445,21 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
                 iobbs(:, [3 4]) = iobbs(:, [3 4]) - iobbs(:, [1 2]) + 1;
                 ioscores = scores(image_only, :);
 
-
+                numWindows = size(iobbs, 1);
                 [iobbs, ioscores, idx] = reduce_matches(params, iobbs, ioscores);
                 info('Reduced %d patches to %d', length(image_only), size(iobbs, 1));
                 for pi=1:length(ioscores)
                     si = image_only(idx(pi));
                     bbs = iobbs(pi, :);
                     bbs([3 4]) = bbs([3 4]) + bbs([1 2]) - 1;
-                    %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
-                    blend_mask = true([size(I, 1) size(I, 2)]);
-                    blend_mask(bbs(2):bbs(4), bbs(1):bbs(3)) = false;
-                    overlay = zeros([size(I, 1) size(I, 2)], 'uint8');
-                    I2 = alpha_blend(I, overlay, 0.4, blend_mask);
-                    %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
-
                     filename = sprintf('%s/%05d-%.3f-Image%d-Patch%d.jpg', target_dir, si, ioscores(pi), image, pi);
-                    imwrite(I2, filename);
+                    if ~params.no_create
+                        %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
+                        I2 = image_with_overlay(I, bbs);
+                        %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
+
+                        imwrite(I2, filename);
+                    end
 
                     result(si).query_curid = model.curid;
                     result(si).curid = database(image).curid;
@@ -486,6 +468,7 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
                     result(si).score = ioscores(pi);
                     result(si).bbox = bbs;
                     result(si).filename = filename;
+                    result(si).num_windows = numWindows;
                 end
             end
         else
@@ -494,6 +477,7 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
             images2 = [];
             resnum = [];
             patches = [];
+            numWindows = [];
             for ii=1:length(umimg)
                 image = umimg(ii);
 
@@ -509,7 +493,6 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
                 iobbs(:, [3 4]) = iobbs(:, [3 4]) - iobbs(:, [1 2]) + 1;
                 ioscores = scores(image_only, :);
 
-
                 [iobbs, ioscores, idx] = reduce_matches(params, iobbs, ioscores);
                 info('Reduced %d patches to %d', length(image_only), size(iobbs, 1));
                 scores2 = [scores2; ioscores];
@@ -517,6 +500,7 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
                 images2 = [images2, ones([1 size(iobbs, 1)]) * image];
                 resnum = [resnum, image_only(idx)];
                 patches = [patches, 1:length(ioscores)];
+                numWindows = [numWindows, ones([1 size(iobbs, 1)]) * size(iobbs, 1)];
             end
             [scores2, idx] = sort(scores2, 'descend');
 
@@ -528,23 +512,32 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
             images2 = images2(idx);
             resnum = resnum(idx);
             patches = patches(idx);
+            numWindows = numWindows(idx);
 
+            img_cache = cell([1 length(database)]);
             for i=1:length(resnum)
+                info('Saving Image %4d/%04d', i, length(resnum), true, true, false);
                 si = resnum(i);
                 pi = patches(i);
                 image = images2(i);
-                I = get_image(params, database(image).curid);
                 bbs = bbs2(i, :);
                 bbs([3 4]) = bbs([3 4]) + bbs([1 2]) - 1;
-                %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
-                blend_mask = true([size(I, 1) size(I, 2)]);
-                blend_mask(bbs(2):bbs(4), bbs(1):bbs(3)) = false;
-                overlay = zeros([size(I, 1) size(I, 2)], 'uint8');
-                I2 = alpha_blend(I, overlay, 0.4, blend_mask);
-                %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
 
                 filename = sprintf('%s/%05d-%.3f-Image%d-Patch%d.jpg', target_dir, si, scores2(i), image, pi);
-                imwrite(I2, filename);
+                if ~params.no_create
+                    fileid = database(image).curid;
+                    if isempty(img_cache{image})
+                        I = get_image(params, fileid);
+                        img_cache{image} = I;
+                    else
+                        I = img_cache{image};
+                    end
+                    %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
+                    I2 = image_with_overlay(I, bbs);
+                    %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
+
+                    imwrite(I2, filename);
+                end
 
                 result(si).query_curid = model.curid;
                 result(si).curid = database(image).curid;
@@ -553,7 +546,9 @@ function results = searchDatabase(params, database, svm_models, fit_params, pos)
                 result(si).score = scores2(i);
                 result(si).bbox = bbs;
                 result(si).filename = filename;
+                result(si).num_windows = numWindows(i);
             end
+            info('', false, true);
         end
 
 %         patchnrs = zeros([max(mimg) 1]);
@@ -645,7 +640,7 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
         mbbs = round(mbbs(idx, :));
         result = struct;
         if ~isempty(scores)
-            result = alloc_struct_array(length(scores), 'curid', 'query_curid', 'img', 'patch', 'score', 'bbox', 'filename');
+            result = alloc_struct_array(length(scores), 'curid', 'query_curid', 'img', 'patch', 'score', 'bbox', 'filename', 'num_windows');
             result(length(scores)).query_curid = model.curid;
         end
 
@@ -665,22 +660,22 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
                 iobbs(:, [3 4]) = iobbs(:, [3 4]) - iobbs(:, [1 2]) + 1;
                 ioscores = scores(image_only, :);
 
-
+                numWindows = size(iobbs, 1);
                 [iobbs, ioscores, idx] = reduce_matches(params, iobbs, ioscores);
                 info('Reduced %d patches to %d', length(image_only), size(iobbs, 1));
                 for pi=1:length(ioscores)
                     si = image_only(idx(pi));
                     bbs = iobbs(pi, :);
                     bbs([3 4]) = bbs([3 4]) + bbs([1 2]) - 1;
-                    %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
-                    blend_mask = true([size(I, 1) size(I, 2)]);
-                    blend_mask(bbs(2):bbs(4), bbs(1):bbs(3)) = false;
-                    overlay = zeros([size(I, 1) size(I, 2)], 'uint8');
-                    I2 = alpha_blend(I, overlay, 0.4, blend_mask);
-                    %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
 
                     filename = sprintf('%s/%05d-%.3f-Image%d-Patch%d.jpg', target_dir, si, ioscores(pi), image, pi);
-                    imwrite(I2, filename);
+                    if ~params.no_create
+                        %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
+                        I2 = image_with_overlay(I, bbs);
+                        %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
+
+                        imwrite(I2, filename);
+                    end
 
                     result(si).query_curid = model.curid;
                     result(si).curid = database(image).curid;
@@ -689,6 +684,7 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
                     result(si).score = ioscores(pi);
                     result(si).bbox = bbs;
                     result(si).filename = filename;
+                    result(si).num_windows = numWindows;
                 end
             end
         else
@@ -697,6 +693,7 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
             images2 = [];
             resnum = [];
             patches = [];
+            numWindows = [];
             for ii=1:length(umimg)
                 image = umimg(ii);
 
@@ -720,6 +717,7 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
                 images2 = [images2, ones([1 size(iobbs, 1)]) * image];
                 resnum = [resnum; image_only(idx)];
                 patches = [patches, 1:length(ioscores)];
+                numWindows = [numWindows, ones([1 size(iobbs, 1)]) * size(iobbs, 1)];
             end
             [scores2, idx] = sort(scores2, 'descend');
 
@@ -731,23 +729,24 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
             images2 = images2(idx);
             resnum = resnum(idx);
             patches = patches(idx);
+            numWindows = numWindows(idx);
 
             for i=1:length(resnum)
                 si = resnum(i);
                 pi = patches(i);
                 image = images2(i);
-                I = get_image(params, database(image).curid);
                 bbs = bbs2(i, :);
                 bbs([3 4]) = bbs([3 4]) + bbs([1 2]) - 1;
-                %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
-                blend_mask = true([size(I, 1) size(I, 2)]);
-                blend_mask(bbs(2):bbs(4), bbs(1):bbs(3)) = false;
-                overlay = zeros([size(I, 1) size(I, 2)], 'uint8');
-                I2 = alpha_blend(I, overlay, 0.4, blend_mask);
-                %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
-
                 filename = sprintf('%s/%05d-%.3f-Image%d-Patch%d.jpg', target_dir, si, scores2(i), image, pi);
-                imwrite(I2, filename);
+
+                if ~params.no_create
+                    I = get_image(params, database(image).curid);
+                    %I2 = I(bbs(2):bbs(4), bbs(1):bbs(3), :);
+                    I2 = image_with_overlay(I, bbs);
+                    %I2 = I(bbs(1):bbs(3), bbs(2):bbs(4), :);
+
+                    imwrite(I2, filename);
+                end
 
                 result(si).query_curid = model.curid;
                 result(si).curid = database(image).curid;
@@ -756,6 +755,7 @@ function results = searchWindowDatabase(params, database, svm_models, fit_params
                 result(si).score = scores2(i);
                 result(si).bbox = bbs;
                 result(si).filename = filename;
+                result(si).num_windows = numWindows(i);
             end
         end
 
@@ -930,7 +930,15 @@ function target_dir = get_target_dir(params, curid)
         filter = 'no';
     end
 
+    if params.window_generation_relative_move > 0
+        sliding_window = sprintf('%.1f', params.window_generation_relative_move);
+    else
+        sliding_window = 'v1';
+    end
+
     target_dir = [params.dataset.localdir filesep 'queries' filesep 'scaled'...
+                filesep sliding_window '-Windows'...
+                filesep sprintf('%.1f', params.max_window_image_ratio) '-WinRatio'...
                 filesep filter '-Filter'...
                 filesep scoring '-Scoring'...
                 filesep int_backend '-Backend'...
